Solve numpy-zeros-ones

+++ study-plans/numpy-basics/numpy-zeros-ones/numpy-zeros-ones.py
@@ -7,4 +7,5 @@ def create_filled_array(shape, kind):
     if kind == 'zeros':
         return np.zeros(shape, dtype = np.float64)
     else:
-        return np.ones(shape, dtype = np.float64)
+        return np.ones(shape)
+
